Bonuses on weapons work fully in combat.

+++ PilotSimulator/Assets/Scripts/ForBattle!/Mods/Bonuses/Bonus.cs
@@ -4,7 +4,7 @@
 public class Bonus : IModData
 {
     [SerializeField] public string type;
-    [SerializeField] public int value;
+    [SerializeField] public float value;
 
     public string ModType { get => type; }
     public float ModValue { get => value; }
